test(mcp-adapters): consolidate schema and hook coverage
